+++ src/Altinn.Notifications.Core/JsonSerializerOptionsProvider.cs
@@ -15,7 +15,8 @@ namespace Altinn.Notifications.Core
         {
             DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
             PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            Converters = { new JsonStringEnumConverter() }
+            Converters = { new JsonStringEnumConverter() },
+            PropertyNameCaseInsensitive = true
         };
     }
 }
