+++ test/Altinn.Notifications.Tests/DelegatingHandlerStub.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Net;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Altinn.Notifications.IntegrationTests
+{
+    public class DelegatingHandlerStub : DelegatingHandler
+    {
+        private readonly Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> _handlerFunc;
+
+        public DelegatingHandlerStub()
+        {
+            _handlerFunc = (request, cancellationToken) => Task.FromResult(new HttpResponseMessage(HttpStatusCode.OK));
+        }
+
+        public DelegatingHandlerStub(Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> handlerFunc)
+        {
+            _handlerFunc = handlerFunc;
+        }
+
+        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            return _handlerFunc(request, cancellationToken);
+        }
+    }
+}
